Flip to new randomness generation backend and remove the existing
