+++ crypto/fipsmodule/bcm.c
@@ -34,10 +34,7 @@
 #pragma bss_seg(".fipsbs$b")
 #endif
 
-#include "rand/new_rand.c"
-#include "rand/entropy/entropy_sources.c"
-#include "rand/entropy/tree_drbg_jitter_entropy.c"
-
+#include <openssl/chacha.h>
 #include <openssl/digest.h>
 #include <openssl/hmac.h>
 #include <openssl/sha.h>
@@ -141,6 +138,8 @@
 #include "rand/ctrdrbg.c"
 #include "rand/rand.c"
 #include "rand/urandom.c"
+#include "rand/entropy/entropy_sources.c"
+#include "rand/entropy/tree_drbg_jitter_entropy.c"
 #include "rsa/blinding.c"
 #include "rsa/padding.c"
 #include "rsa/rsa.c"
